mock-variables.c: fix gcc warning
$ make test

    ...
    mock-variables.c: In function 'mock_sv_attrs_match':
    mock-variables.c:448:23: error: conversion from 'long unsigned int' to 'UINT32'
        {aka 'unsigned int'} changes value from '18446744073709551551' to '4294967231' [-Werror=overflow]
      448 |         UINT32 mask = ~EFI_VARIABLE_APPEND_WRITE;
          |                       ^
    cc1: all warnings being treated as errors

+++ mock-variables.c
@@ -445,7 +445,7 @@ free_var(struct mock_variable *var)
 static bool
 mock_sv_attrs_match(UINT32 old, UINT32 new)
 {
-	UINT32 mask = ~EFI_VARIABLE_APPEND_WRITE;
+	UINT32 mask = ~((UINT32)EFI_VARIABLE_APPEND_WRITE);
 
 	return (old & mask) == (new & mask);
 }
